refactor: move param properties to its own panel
