+++ src/screens/surrealist/views/parameters/ParameterPropertiesPanel/style.module.scss
